+++ report_qweb_field_options/README.rst
@@ -45,10 +45,10 @@ For each record:
 - Set **Model** and **Field** (required)
 - Set **UoM** and **UoM Field**, or **Currency** and **Currency Field** only for fields
   of float type (optional)
-- Set **Options** as a string representation of a dictionary (e.g., ``{"widget": "date"}``,
-  ``{"widget": "monetary"}``, or ``{'widget': 'contact', 'fields': ['name', 'phone']}``)
 - Set **Company** (optional)
-- Set **Digits** (required only for float-type fields)
+- Set **Options** as a string representation of a dictionary. E.g., ``{"widget": "date"}``,
+  ``{"widget": "monetary"}``, or ``{'widget': 'contact', 'fields': ['name', 'phone']}``
+- Set **Digits** (only for float-type fields). The value is ignored if Options is set
 
 Usage
 =====
