+++ report_qweb_field_options/models/__init__.py
@@ -1,3 +1,2 @@
 from . import ir_qweb
-from . import ir_qweb_fields
 from . import qweb_field_options
