+++ report_qweb_field_options/static/description/index.html
@@ -395,10 +395,10 @@ according to your needs.</p>
 <li>Set <strong>Model</strong> and <strong>Field</strong> (required)</li>
 <li>Set <strong>UoM</strong> and <strong>UoM Field</strong>, or <strong>Currency</strong> and <strong>Currency Field</strong> only for fields
 of float type (optional)</li>
-<li>Set <strong>Options</strong> as a string representation of a dictionary (e.g., <tt class="docutils literal">{&quot;widget&quot;: &quot;date&quot;}</tt>,
-<tt class="docutils literal">{&quot;widget&quot;: &quot;monetary&quot;}</tt>, or <tt class="docutils literal">{'widget': 'contact', 'fields': ['name', <span class="pre">'phone']}</span></tt>)</li>
 <li>Set <strong>Company</strong> (optional)</li>
-<li>Set <strong>Digits</strong> (required only for float-type fields)</li>
+<li>Set <strong>Options</strong> as a string representation of a dictionary. E.g., <tt class="docutils literal">{&quot;widget&quot;: &quot;date&quot;}</tt>,
+<tt class="docutils literal">{&quot;widget&quot;: &quot;monetary&quot;}</tt>, or <tt class="docutils literal">{'widget': 'contact', 'fields': ['name', <span class="pre">'phone']}</span></tt></li>
+<li>Set <strong>Digits</strong> (only for float-type fields). The value is ignored if Options is set</li>
 </ul>
 </div>
 <div class="section" id="usage">
